pyproject.toml
[tool.poetry]
name = "nextjs-lesson-rest-api"
version = "0.1.0"
description = ""
authors = ["Yuya Shintani <yuya.shintani.uplus26e7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
django-cors-headers = "^3.6.0"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"
djoser = "^2.1.0"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
dj-static = "^0.0.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
